+++ packages/driver/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- fix capabilities parsing for native apps
 
 ## 1.3.3 - 2021/11/14